# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GAMEA Social Monitor",
    "description": "API Institucional de Monitoreo y Analítica de Interacciones en Redes Sociales del Gobierno Autónomo Municipal de El Alto (GAMEA).",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/monitoring/hub/config": {
      "get": {
        "tags": [
          "Monitoring & Social Scraper Hub"
        ],
        "summary": "Obtener configuración de conectores de redes sociales (Facebook & TikTok)",
        "operationId": "get_connector_configs_api_v1_monitoring_hub_config_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectorConfigsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Monitoring & Social Scraper Hub"
        ],
        "summary": "Actualizar credenciales y parámetros de extracción de Facebook y TikTok",
        "operationId": "update_connector_configs_api_v1_monitoring_hub_config_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectorConfigUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectorConfigsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/monitoring/hub/test-connection": {
      "post": {
        "tags": [
          "Monitoring & Social Scraper Hub"
        ],
        "summary": "Probar conectividad y validar credenciales de Facebook o TikTok",
        "operationId": "test_connection_api_v1_monitoring_hub_test_connection_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestConnectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestConnectionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/monitoring/hub/diagnostics": {
      "get": {
        "tags": [
          "Monitoring & Social Scraper Hub"
        ],
        "summary": "Diagnóstico exhaustivo de variables de entorno y estado de conectores API",
        "operationId": "get_connectors_diagnostics_api_v1_monitoring_hub_diagnostics_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectorsDiagnosticResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/monitoring/hub/audience": {
      "get": {
        "tags": [
          "Monitoring & Social Scraper Hub"
        ],
        "summary": "Listar personas en monitoreo con sus perfiles de Facebook y TikTok",
        "operationId": "get_audience_api_v1_monitoring_hub_audience_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MonitoredPersonResponse"
                  },
                  "type": "array",
                  "title": "Response Get Audience Api V1 Monitoring Hub Audience Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Monitoring & Social Scraper Hub"
        ],
        "summary": "Adicionar o actualizar una persona en la lista de monitoreo",
        "operationId": "add_monitored_person_api_v1_monitoring_hub_audience_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MonitoredPersonCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonitoredPersonResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/monitoring/hub/audience/bulk": {
      "post": {
        "tags": [
          "Monitoring & Social Scraper Hub"
        ],
        "summary": "Importación masiva de personas a monitorear (CSV o JSON)",
        "operationId": "bulk_import_audience_api_v1_monitoring_hub_audience_bulk_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MonitoredPersonBulkImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonitoredPersonBulkImportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/monitoring/hub/sync-now": {
      "post": {
        "tags": [
          "Monitoring & Social Scraper Hub"
        ],
        "summary": "Ejecutar extracción/scrapeo y cruce de interacciones en Facebook y TikTok",
        "operationId": "run_social_sync_api_v1_monitoring_hub_sync_now_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunSyncRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunSyncResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/monitoring/hub/activity-matrix": {
      "get": {
        "tags": [
          "Monitoring & Social Scraper Hub"
        ],
        "summary": "Consultar matriz cruzada de reacciones y actividades de la audiencia",
        "operationId": "get_activity_matrix_api_v1_monitoring_hub_activity_matrix_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "publication_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtrar por publicación específica",
              "title": "Publication Id"
            },
            "description": "Filtrar por publicación específica"
          },
          {
            "name": "platform",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtrar por plataforma (ALL, FACEBOOK, TIKTOK)",
              "title": "Platform"
            },
            "description": "Filtrar por plataforma (ALL, FACEBOOK, TIKTOK)"
          },
          {
            "name": "department",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtrar por unidad organizacional",
              "title": "Department"
            },
            "description": "Filtrar por unidad organizacional"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Búsqueda por nombre, CI o usuario",
              "title": "Search"
            },
            "description": "Búsqueda por nombre, CI o usuario"
          },
          {
            "name": "participation_status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ALL, PARTICIPATED, NO_ACTIVITY",
              "default": "ALL",
              "title": "Participation Status"
            },
            "description": "ALL, PARTICIPATED, NO_ACTIVITY"
          },
          {
            "name": "max_posts",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Ventana de publicaciones recientes a evaluar (las verificadas siempre se incluyen)",
              "default": 15,
              "title": "Max Posts"
            },
            "description": "Ventana de publicaciones recientes a evaluar (las verificadas siempre se incluyen)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityMatrixResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/monitoring/hub/verify-employee-activity": {
      "post": {
        "tags": [
          "Monitoring & Social Scraper Hub"
        ],
        "summary": "Registrar o auditar manualmente la interacción de un funcionario",
        "operationId": "verify_employee_activity_api_v1_monitoring_hub_verify_employee_activity_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmployeeActivityVerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeActivityVerifyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/monitoring/hub/import-reactions": {
      "post": {
        "tags": [
          "Monitoring & Social Scraper Hub"
        ],
        "summary": "Importación y cruce masivo de reacciones de Facebook contra el padrón municipal",
        "operationId": "import_reactions_batch_api_v1_monitoring_hub_import_reactions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportReactionsBatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportReactionsBatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/monitoring/hub/export-matrix": {
      "get": {
        "tags": [
          "Monitoring & Social Scraper Hub"
        ],
        "summary": "Descargar informe oficial de fiscalización en formato Excel (.xlsx)",
        "operationId": "export_matrix_excel_api_v1_monitoring_hub_export_matrix_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "publication_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Publication Id"
            }
          },
          {
            "name": "platform",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Platform"
            }
          },
          {
            "name": "department",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Department"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "participation_status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "ALL",
              "title": "Participation Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConnectorConfigUpdateRequest": {
        "properties": {
          "configs": {
            "items": {
              "$ref": "#/components/schemas/ConnectorConfigItem"
            },
            "type": "array",
            "title": "Configs"
          }
        },
        "type": "object",
        "required": [
          "configs"
        ],
        "title": "ConnectorConfigUpdateRequest"
      },
      "ConnectorConfigsResponse": {
        "properties": {
          "configs": {
            "items": {
              "$ref": "#/components/schemas/ConnectorConfigItem"
            },
            "type": "array",
            "title": "Configs"
          }
        },
        "type": "object",
        "required": [
          "configs"
        ],
        "title": "ConnectorConfigsResponse"
      },
...
      "ConnectorsDiagnosticResponse": {
        "properties": {
          "connectors": {
            "items": {
              "$ref": "#/components/schemas/ConnectorDiagnosticDetail"
            },
            "type": "array",
            "title": "Connectors"
          },
          "system_env": {
            "type": "string",
            "title": "System Env"
          },
          "all_operational": {
            "type": "boolean",
            "title": "All Operational"
          },
          "total_missing_variables": {
            "type": "integer",
            "title": "Total Missing Variables"
          }
        },
        "type": "object",
        "required": [
          "connectors",
          "system_env",
          "all_operational",
          "total_missing_variables"
        ],
        "title": "ConnectorsDiagnosticResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MonitoredPersonResponse": {
        "properties": {
          "ci": {
            "type": "string",
            "title": "Ci"
          },
          "first_name": {
            "type": "string",
            "title": "First Name"
          },
          "last_name": {
            "type": "string",
            "title": "Last Name"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "department": {
            "type": "string",
            "title": "Department"
          },
          "position": {
            "type": "string",
            "title": "Position"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "facebook_account": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Facebook Account"
          },
          "facebook_profile_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Facebook Profile Url"
          },
          "tiktok_account": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tiktok Account"
          },
          "tiktok_profile_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tiktok Profile Url"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "ci",
          "first_name",
          "last_name",
          "full_name",
          "department",
          "position",
          "email",
          "facebook_account",
          "facebook_profile_url",
          "tiktok_account",
          "tiktok_profile_url",
          "status",
          "created_at"
        ],
        "title": "MonitoredPersonResponse"
      },
...
      "TestConnectionRequest": {
        "properties": {
          "platform_name": {
            "type": "string",
            "title": "Platform Name"
          },
          "target_account_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Account Id"
          },
          "access_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Access Token"
          },
          "api_secret": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Secret"
          },
          "extraction_mode": {
            "type": "string",
            "title": "Extraction Mode",
            "default": "OFFICIAL_API"
          }
        },
        "type": "object",
        "required": [
          "platform_name"
        ],
        "title": "TestConnectionRequest"
      },
      "TestConnectionResponse": {
        "properties": {
          "platform_name": {
            "type": "string",
            "title": "Platform Name"
          },
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "account_info": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Info"
          }
        },
        "type": "object",
        "required": [
          "platform_name",
          "success",
          "status",
          "message"
        ],
        "title": "TestConnectionResponse"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}